added new jar file to tools/ dir
